requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
PyMuPDF==1.24.3
python-docx==1.1.0
spacy==3.7.4
keybert==0.8.4
sentence-transformers==2.7.0
transformers==4.41.0
torch==2.3.0
nltk==3.8.1
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0